pyproject.toml
[project]
name = "pydantic-ai-skills"
version = "0.1.0"
description = "Claude Code skills for building and evaluating Pydantic AI agents"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = [
    "pydantic",
    "pydantic-ai",
    "ai-agents",
    "llm",
    "evaluation",
    "claude-code",
    "skills"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "pydantic-ai>=1.44.0",
    "pydantic-evals>=1.44.0",
    "logfire>=4.17.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.13",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["skills", "exampleCode"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.13",
    "mypy>=1.10.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
markers = [
    "real_api: Tests requiring real API access (skipped without API keys)",
    "slow: Tests that take a long time to run",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["skills", "exampleCode"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
